Conflicts:
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810011.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810012.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810013.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810014.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810015.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810021.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810022.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810023.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810024.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810025.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810026.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810027.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810028.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810029.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810030.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810031.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0810032.raml
	DB/DBC/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBC0820024.raml
	DB/DBE/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBE2020009.raml
	DB/DBE/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBE2200001.raml
	DB/DBE/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBE2300001.raml
	DB/DBE/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBE2330001.raml
	DB/DBE/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBE5130001.raml
	DB/DBE/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBE5140001.raml
	DB/DBE/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBE5140002.raml
	DB/DBE/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBE5140003.raml
	DB/DBE/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBE5250002.raml
	DB/DBU/standard/divcontroller/src/main/java/jp/co/ndensan/reams/db/dbu/divcontroller/entity/parentdiv/DBU0020021/DBU0020021StateName.java
	DB/DBU/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBU0020011.raml
	DB/DBU/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBU0020021.raml
	DB/DBU/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBU0020051.raml
	DB/DBU/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBU0020091.raml
	DB/DBU/standard/divcontroller/src/main/webapp/WEB-INF/standard/View/DBU0060031.raml

+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/ShichosonTokubetsuKyufuJigyoshaBuilder.java
@@ -73,7 +73,7 @@ public class ShichosonTokubetsuKyufuJigyoshaBuilder {
      * @param 履歴番号 履歴番号
      * @return {@link ShichosonTokubetsuKyufuJigyoshaBuilder}
      */
-    public ShichosonTokubetsuKyufuJigyoshaBuilder set履歴番号(Decimal 履歴番号) {
+    public ShichosonTokubetsuKyufuJigyoshaBuilder set履歴番号(int 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         entity.setRirekiNo(履歴番号);
         return this;
