+++ DB/DBC/standard/business/src/test/java/jp/co/ndensan/reams/db/dbc/business/core/basic/ShichosonTokubetsuKyufuJigyoshaIdentifierTest.java
@@ -27,7 +27,7 @@ public class ShichosonTokubetsuKyufuJigyoshaIdentifierTest extends DbcTestBase {
 //TODO 主キーの数が足りない場合、追加してください。
     private static JigyoshaNo 主キー名1;
     private static ServiceCode 主キー名2;
-    private static Decimal 主キー名3;
+    private static int 主キー名3;
 
     @BeforeClass
     public static void setUpClass() {
